+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-# 0.1.0
+# Change Log
 
-Initial release of rbenv
+## [unreleased] - unreleased
+### Added
+Initial release of rbenv. Now you can
+- install rbenv command for a single user
+- install a ruby version using rbenv
+- install a gem for a ruby version using rbenv